+++ client/src/pages/_app.tsx
@@ -1,0 +1,21 @@
+import type { AppProps } from 'next/app';
+import { ApolloProvider } from '@apollo/client';
+import { ApolloClient, InMemoryCache, HttpLink } from '@apollo/client';
+
+// Create Apollo Client
+const client = new ApolloClient({
+  ssrMode: typeof window === 'undefined',
+  link: new HttpLink({
+    uri: process.env.NEXT_PUBLIC_GRAPHQL_URL || 'http://localhost:4000/graphql',
+    credentials: 'include',
+  }),
+  cache: new InMemoryCache(),
+});
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <ApolloProvider client={client}>
+      <Component {...pageProps} />
+    </ApolloProvider>
+  );
+}